events verifyToken now working

+++ controllers/events.js
@@ -5,13 +5,13 @@ const router = express.Router();
 const { Event } = require("../models");
 
 // Add all the CRUD features bellow
-// router.use(verifyToken);
+router.use(verifyToken);
 
 router.post("/", async (req, res) => {
   try {
-    // req.body.author = req.user._id;
+    req.body.author = req.user._id;
     const event = await Event.create(req.body);
-    // event._doc.author = req.user;
+    event._doc.author = req.user;
     res.status(201).json(event);
   } catch (error) {
     console.log(error);
@@ -45,9 +45,9 @@ router.put("/:eventId", async (req, res) => {
   try {
     const event = await Event.findById(req.params.eventId);
 
-    // if (!event.author.equals(req.user._id)) {
-    //   return res.status(403).send(`You're not allowed to do that!`);
-    // }
+    if (!event.author.equals(req.user._id)) {
+      return res.status(403).send(`You're not allowed to do that!`);
+    }
 
     const updatedEvent = await Event.findByIdAndUpdate(
       req.params.eventId,
@@ -55,7 +55,7 @@ router.put("/:eventId", async (req, res) => {
       { new: true }
     );
 
-    // updatedEvent._doc.author = req.user;
+    updatedEvent._doc.author = req.user;
 
     res.status(200).json(updatedEvent);
   } catch (error) {
@@ -67,9 +67,9 @@ router.delete("/:eventId", async (req, res) => {
   try {
     const event = await Event.findById(req.params.eventId);
 
-    // if (!event.author.equals(req.user._id)) {
-    //   return res.status(403).send(`You're not allowed to do that!`);
-    // }
+    if (!event.author.equals(req.user._id)) {
+      return res.status(403).send(`You're not allowed to do that!`);
+    }
 
     const deletedEvent = await Event.findByIdAndDelete(req.params.eventId);
     res.status(200).json(deletedEvent);
